Adding remixcam.renderers package. It remains to simplify the API, create proscene.renderers package and rename the classes after that.

+++ src/remixlab/remixcam/devices/DeviceGrabbable.java
@@ -140,5 +140,5 @@ public interface DeviceGrabbable {
 	 * Callback method called when the MouseGrabber {@link #grabsMouse()} and the
 	 * mouse wheel is used.
 	 */
-	void mouseWheelMoved(int rotation, Pinhole vp);
+	void mouseWheelMoved(float rotation, Pinhole vp);
 }
